requirements.txt

Flask==3.0.3
databricks-sql-connector==3.2.0
python-dotenv==1.0.1
Flask-Cors==4.0.1
numpy==1.26.4
pandas==2.1.4
gunicorn==22.0.0
pytest==8.2.2
pytest-mock==3.14.0
